Unit Component 업데이트 전투로직 추가

+++ Assets/Scripts/Unit/Stat/UnitStatData.cs
@@ -6,7 +6,7 @@ public class UnitStatData
     public int UnitId { get; }
     public int StarLevel { get; }
     public int BaseHp { get; }
-    public float BaseAtk { get; }
+    public int BaseAtk { get; }
 
     public UnitStatData(DB_UnitStats e)
     {
